+++ src/smoother.h
@@ -15,14 +15,14 @@ class Smoother
 
   private:
     std::vector<float> _rbuf; // rotary buffer
-    int _kernelsize;          // filter kernel width
-    int _rotater;             //pointer to current sample in rotary buffer
+    uint16_t _kernelsize;          // filter kernel width
+    uint16_t _rotater;             //pointer to current sample in rotary buffer
     float _runner;            // current filter output value
     bool _ready = false;
 
   public:
-    void init(int width);
-    void init(int width, float value);
+    void init(uint16_t width);
+    void init(uint16_t width, float value);
     float addSample(float sample);
     float latest();
     void reset(void);
